broken before

+++ backport-bin.sh
@@ -18,12 +18,11 @@ fi
 
 cat $filename | python3 pre-dose.py $2_Packages $3_Packages > modified_Packages
 
-dose-debcheck --deb-native-arch=amd64 -e -f $3_Packages \
-    | grep unsat-dep | awk '{print $2}' | cut -f 1 -d ":" | sort -u > $base_name.broken.before
-
-cat $base_name.broken.before \
-        | sort -u | python3 pre-dose.py -d $2_Packages modified_Packages > modified_Packages.tmp && \
-        mv -f modified_Packages.tmp modified_Packages    
+if [ -e $3_Packages.broken.before ]; then
+    cat $3_Packages.broken.before \
+            | sort -u | python3 pre-dose.py -d $2_Packages modified_Packages > modified_Packages.tmp && \
+            mv -f modified_Packages.tmp modified_Packages    
+fi
 
 while [ -s "$filename" ]; do
     echo "Processing $filename"
